+++ src/test/java/com/hyperwallet/clientsdk/HyperwalletIT.java
@@ -396,10 +396,10 @@ public class HyperwalletIT {
         assertThat(returnValue.getCreatedOn(), is(equalTo(dateFormat.parse("2017-10-31T22:32:57 UTC"))));
         assertThat(returnValue.getTransferMethodCountry(), is(equalTo("US")));
         assertThat(returnValue.getTransferMethodCurrency(), is(equalTo("USD")));
-        assertThat(returnValue.getBankAccountRelationship(), is(equalTo(HyperwalletTransferMethod.BankAccountRelationship.SELF)));
-        assertThat(returnValue.getProfileType(), is(equalTo(HyperwalletUser.ProfileType.INDIVIDUAL)));
+        assertThat(returnValue.getProfileType(), is(equalTo(HyperwalletUser.ProfileType.BUSINESS)));
         assertThat(returnValue.getFirstName(), is(equalTo("Some")));
         assertThat(returnValue.getLastName(), is(equalTo("Guy")));
+        assertThat(returnValue.getBusinessOperatingName(), is(equalTo("ABC Framing")));
         assertThat(returnValue.getDateOfBirth(), is(equalTo(dateFormat.parse("1991-01-01T00:00:00 UTC"))));
         assertThat(returnValue.getAddressLine1(), is(equalTo("575 Market Street")));
         assertThat(returnValue.getCity(), is(equalTo("San Francisco")));
@@ -434,10 +434,10 @@ public class HyperwalletIT {
         assertThat(returnValue.getCreatedOn(), is(equalTo(dateFormat.parse("2017-10-31T22:42:46 UTC"))));
         assertThat(returnValue.getTransferMethodCountry(), is(equalTo("US")));
         assertThat(returnValue.getTransferMethodCurrency(), is(equalTo("USD")));
-        assertThat(returnValue.getBankAccountRelationship(), is(equalTo(HyperwalletTransferMethod.BankAccountRelationship.SELF)));
-        assertThat(returnValue.getProfileType(), is(equalTo(HyperwalletUser.ProfileType.INDIVIDUAL)));
+        assertThat(returnValue.getProfileType(), is(equalTo(HyperwalletUser.ProfileType.BUSINESS)));
         assertThat(returnValue.getFirstName(), is(equalTo("Some")));
         assertThat(returnValue.getLastName(), is(equalTo("Guy")));
+        assertThat(returnValue.getBusinessOperatingName(), is(equalTo("ABC Framing")));
         assertThat(returnValue.getDateOfBirth(), is(equalTo(dateFormat.parse("1991-01-01T00:00:00 UTC"))));
         assertThat(returnValue.getAddressLine1(), is(equalTo("123 Other Street")));
         assertThat(returnValue.getCity(), is(equalTo("Far Away Land")));
@@ -467,10 +467,10 @@ public class HyperwalletIT {
         assertThat(returnValue.getCreatedOn(), is(equalTo(dateFormat.parse("2017-10-31T22:32:57 UTC"))));
         assertThat(returnValue.getTransferMethodCountry(), is(equalTo("US")));
         assertThat(returnValue.getTransferMethodCurrency(), is(equalTo("USD")));
-        assertThat(returnValue.getBankAccountRelationship(), is(equalTo(HyperwalletTransferMethod.BankAccountRelationship.SELF)));
-        assertThat(returnValue.getProfileType(), is(equalTo(HyperwalletUser.ProfileType.INDIVIDUAL)));
+        assertThat(returnValue.getProfileType(), is(equalTo(HyperwalletUser.ProfileType.BUSINESS)));
         assertThat(returnValue.getFirstName(), is(equalTo("Some")));
         assertThat(returnValue.getLastName(), is(equalTo("Guy")));
+        assertThat(returnValue.getBusinessOperatingName(), is(equalTo("ABC Framing")));
         assertThat(returnValue.getDateOfBirth(), is(equalTo(dateFormat.parse("1991-01-01T00:00:00 UTC"))));
         assertThat(returnValue.getAddressLine1(), is(equalTo("575 Market Street")));
         assertThat(returnValue.getCity(), is(equalTo("San Francisco")));
@@ -500,10 +500,10 @@ public class HyperwalletIT {
         assertThat(returnValue.getData().get(0).getCreatedOn(), is(equalTo(dateFormat.parse("2017-10-31T22:32:57 UTC"))));
         assertThat(returnValue.getData().get(0).getTransferMethodCountry(), is(equalTo("US")));
         assertThat(returnValue.getData().get(0).getTransferMethodCurrency(), is(equalTo("USD")));
-        assertThat(returnValue.getData().get(0).getBankAccountRelationship(), is(equalTo(HyperwalletTransferMethod.BankAccountRelationship.SELF)));
-        assertThat(returnValue.getData().get(0).getProfileType(), is(equalTo(HyperwalletUser.ProfileType.INDIVIDUAL)));
+        assertThat(returnValue.getData().get(0).getProfileType(), is(equalTo(HyperwalletUser.ProfileType.BUSINESS)));
         assertThat(returnValue.getData().get(0).getFirstName(), is(equalTo("Some")));
         assertThat(returnValue.getData().get(0).getLastName(), is(equalTo("Guy")));
+        assertThat(returnValue.getData().get(0).getBusinessOperatingName(), is(equalTo("ABC Framing")));
         assertThat(returnValue.getData().get(0).getDateOfBirth(), is(equalTo(dateFormat.parse("1991-01-01T00:00:00 UTC"))));
         assertThat(returnValue.getData().get(0).getAddressLine1(), is(equalTo("575 Market Street")));
         assertThat(returnValue.getData().get(0).getCity(), is(equalTo("San Francisco")));
@@ -945,6 +945,105 @@ public class HyperwalletIT {
         assertThat(returnValue.getData().get(0).getTransferMethodCountry(), is(equalTo("US")));
         assertThat(returnValue.getData().get(0).getTransferMethodCurrency(), is(equalTo("USD")));
         assertThat(returnValue.getData().get(0).getEmail(), is(equalTo("user@domain.com")));
+    }
+
+    @Test
+    public void testDeactivatePayPalAccount() throws Exception {
+        String functionality = "deactivatePayPalAccount";
+        initMockServer(functionality);
+
+        HyperwalletStatusTransition returnValue;
+
+        try {
+            returnValue = client.deactivatePayPalAccount("usr-f695ef43-9614-4e17-9269-902c234616c3",
+                    "trm-da21954a-3910-4d70-b83d-0c2d96104394", "PayPal account removed.");
+        } catch (Exception e) {
+            mockServer.verify(parseRequest(functionality));
+            throw e;
+        }
+
+        assertThat(returnValue.getToken(), is(equalTo("sts-70ddc78a-0c14-4a72-8390-75d49ff376f2")));
+        assertThat(returnValue.getCreatedOn(), is(equalTo(dateFormat.parse("2018-10-30T18:50:20 UTC"))));
+        assertThat(returnValue.getTransition(), is(equalTo(DE_ACTIVATED)));
+        assertThat(returnValue.getFromStatus(), is(equalTo(ACTIVATED)));
+        assertThat(returnValue.getToStatus(), is(equalTo(DE_ACTIVATED)));
+        assertThat(returnValue.getNotes(), is(equalTo("PayPal account removed.")));
+    }
+
+    @Test
+    public void testListPayPalAccountStatusTransitions() throws Exception {
+        String functionality = "listPayPalAccountStatusTransitions";
+        initMockServer(functionality);
+
+        HyperwalletList<HyperwalletStatusTransition> returnValue;
+        try {
+            returnValue = client.listPayPalAccountStatusTransitions("usr-f695ef43-9614-4e17-9269-902c234616c3",
+                    "trm-da21954a-3910-4d70-b83d-0c2d96104394");
+        } catch (Exception e) {
+            mockServer.verify(parseRequest(functionality));
+            throw e;
+        }
+
+        assertThat(returnValue.getCount(), is(equalTo(1)));
+        assertThat(returnValue.getOffset(), is(equalTo(0)));
+        assertThat(returnValue.getLimit(), is(equalTo(10)));
+        assertThat(returnValue.getData().get(0).getToken(), is(equalTo("sts-70ddc78a-0c14-4a72-8390-75d49ff376f2")));
+        assertThat(returnValue.getData().get(0).getCreatedOn(), is(equalTo(dateFormat.parse("2018-10-30T18:50:20 UTC"))));
+        assertThat(returnValue.getData().get(0).getTransition(), is(equalTo(DE_ACTIVATED)));
+        assertThat(returnValue.getData().get(0).getFromStatus(), is(equalTo(ACTIVATED)));
+        assertThat(returnValue.getData().get(0).getToStatus(), is(equalTo(DE_ACTIVATED)));
+        assertThat(returnValue.getData().get(0).getNotes(), is(equalTo("PayPal account is removed.")));
+    }
+
+    @Test
+    public void testCreatePayPalAccountStatusTransition() throws Exception {
+        String functionality = "createPayPalAccountStatusTransition";
+        initMockServer(functionality);
+
+        HyperwalletStatusTransition transition = new HyperwalletStatusTransition();
+        transition.setNotes("PayPal account removed.");
+        transition.setTransition(DE_ACTIVATED);
+
+
+        HyperwalletStatusTransition returnValue;
+        try {
+            returnValue = client.createPayPalAccountStatusTransition("usr-f695ef43-9614-4e17-9269-902c234616c3",
+                    "trm-da21954a-3910-4d70-b83d-0c2d96104394",
+                    transition);
+        } catch (Exception e) {
+            mockServer.verify(parseRequest(functionality));
+            throw e;
+        }
+
+        assertThat(returnValue.getToken(), is(equalTo("sts-70ddc78a-0c14-4a72-8390-75d49ff376f2")));
+        assertThat(returnValue.getCreatedOn(), is(equalTo(dateFormat.parse("2018-10-30T18:50:20 UTC"))));
+        assertThat(returnValue.getTransition(), is(equalTo(DE_ACTIVATED)));
+        assertThat(returnValue.getFromStatus(), is(equalTo(ACTIVATED)));
+        assertThat(returnValue.getToStatus(), is(equalTo(DE_ACTIVATED)));
+        assertThat(returnValue.getNotes(), is(equalTo("PayPal account removed.")));
+    }
+
+    @Test
+    public void testGetPayPalAccountStatusTransition() throws Exception {
+        String functionality = "getPayPalAccountStatusTransition";
+        initMockServer(functionality);
+
+        HyperwalletStatusTransition returnValue;
+        try {
+            returnValue = client.getPayPalAccountStatusTransition("usr-f695ef43-9614-4e17-9269-902c234616c3",
+                    "trm-da21954a-3910-4d70-b83d-0c2d96104394",
+                    "sts-70ddc78a-0c14-4a72-8390-75d49ff376f2");
+        } catch (Exception e) {
+            mockServer.verify(parseRequest(functionality));
+            throw e;
+        }
+
+        assertThat(returnValue.getToken(), is(equalTo("sts-70ddc78a-0c14-4a72-8390-75d49ff376f2")));
+        assertThat(returnValue.getCreatedOn(), is(equalTo(dateFormat.parse("2018-10-30T18:50:20 UTC"))));
+        assertThat(returnValue.getTransition(), is(equalTo(DE_ACTIVATED)));
+        assertThat(returnValue.getFromStatus(), is(equalTo(ACTIVATED)));
+        assertThat(returnValue.getToStatus(), is(equalTo(DE_ACTIVATED)));
+        assertThat(returnValue.getNotes(), is(equalTo("PayPal account is removed.")));
     }
 
     //