+++ src/test/java/com/hyperwallet/clientsdk/model/HyperwalletTransferMethodTest.java
@@ -22,7 +22,6 @@ public class HyperwalletTransferMethodTest extends BaseModelTest<HyperwalletTran
                 .branchName("test-branch-name")
                 .branchId("test-branch-id")
                 .bankAccountId("test-bank-account-id")
-                .bankAccountRelationship(HyperwalletTransferMethod.BankAccountRelationship.SELF)
                 .bankAccountPurpose("test-bank-account-purpose")
 
                 .branchAddressLine1("test-branch-address-line1")
@@ -55,6 +54,7 @@ public class HyperwalletTransferMethodTest extends BaseModelTest<HyperwalletTran
                 .profileType(HyperwalletUser.ProfileType.INDIVIDUAL)
 
                 .businessName("test-business-name")
+                .businessOperatingName("test-business-operating-name")
                 .businessRegistrationId("test-business-registration-id")
                 .businessRegistrationStateProvince("test-business-registration-state-province")
                 .businessRegistrationCountry("test-business-registration-country")