+++ CHANGELOG.md
@@ -1,7 +1,22 @@
 Changelog
 =========
+1.4.2
+------------------
+- Added status param to HyperwalletPayment
 
-1.3.0
+1.4.1
+-------------------
+- Added support for new businessType field values. 
+
+1.4.0
+-------------------
+- Fix TypeError thrown when response status is 204 No content
+- Add CVV field to the SDK
+- Remove Relationship field from Server SDK
+- Add Business Operating Name Field to User
+- Add PayPal account status transitions
+
+1.3.0 (2019-01-28)
 -------------------
 - Added field "VerificationStatus" to User
 - Client-token endpoint renamed to authentication-token